+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [linwave] Add `set_range_index` method
 - [linwave] Add Pulsing example for linWave
 - [linwave] Add `identify` method to blind all LEDs or single channel to identify device/channel
+- Add Python 3.11 to CI pipeline
 
 
 ## [0.6.0] - 2022-08-01